% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glyread-package.R
\docType{package}
\name{glyread-package}
\alias{glyread}
\alias{glyread-package}
\title{glyread: Read and process glycomics and glycoproteomics data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package reads and processes quantification results from common glycomics and glycoproteomics software, such as Byonic, StrucGP, or pGlyco, and converts them into a `glyexp_experiment` object (from the [glyexp](https://github.com/glycoverse/glyexp) package) for further analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://glycoverse.github.io/glyread/}
}

}
\author{
\strong{Maintainer}: Bin Fu \email{23110220018@m.fudan.edu.cn} (\href{https://orcid.org/0000-0001-8567-2997}{ORCID})

}
\keyword{internal}
