% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/byonic-byologic.R
\name{read_byonic_byologic}
\alias{read_byonic_byologic}
\title{Read Byonic-Byologic result}
\usage{
read_byonic_byologic(
  fp,
  sample_info = NULL,
  quant_method = "label-free",
  glycan_type = "N",
  sample_name_converter = NULL,
  orgdb = "org.Hs.eg.db"
)
}
\arguments{
\item{fp}{File path of the pGlycoQuant result file.}

\item{sample_info}{File path of the sample information file (csv),
or a sample information data.frame/tibble.}

\item{quant_method}{Quantification method. Either "label-free" or "TMT".}

\item{glycan_type}{Glycan type. Either "N" or "O". Default is "N".}

\item{sample_name_converter}{A function to convert sample names from file paths.
The function should take a character vector of old sample names
and return new sample names.
Note that sample names in \code{sample_info} should match the new names.
If NULL, original names are kept.}

\item{orgdb}{name of the OrgDb package to use for UniProt to gene symbol conversion.
Default is "org.Hs.eg.db".}
}
\value{
An \code{\link[glyexp:experiment]{glyexp::experiment()}} object.
}
\description{
If you used Byonic for intact glycopeptide identification,
and used Byologic for quantification, this is the function for you.
It reads in a result file and returns a \code{\link[glyexp:experiment]{glyexp::experiment()}} object.
Currently only label-free quantification is supported.
}
\section{Which file to use?}{

Open the .blgc file in the result folder with PMI-Byos.
In the "Peptide List" panel (usually on the bottom right),
click "Export content of the table to a CSV file" button.
The exported .csv file is the file you should use.
}

\section{Multisite glycopeptides}{

Multisite glycopeptides are supported but their \code{protein_site} will be set to \code{NA}
since the exact site of glycosylation cannot be determined unambiguously.
}

\section{Variable information}{

The following columns could be found in the variable information tibble:
\itemize{
\item \code{peptide}: character, peptide sequence
\item \code{peptide_site}: integer, site of glycosylation on peptide
\item \code{protein}: character, protein accession
\item \code{protein_site}: integer, site of glycosylation on protein
\item \code{gene}: character, gene name (symbol)
\item \code{glycan_composition}: \code{\link[glyrepr:glycan_composition]{glyrepr::glycan_composition()}}, glycan compositions.
}
}

\section{Sample information}{
The sample information file should be a \code{csv} file with the first column
named \code{sample}, and the rest of the columns being sample information.
The \code{sample} column must match the \code{RawName} column in the pGlyco3 result file,
although the order can be different.

You can put any useful information in the sample information file.
Recommended columns are:
\itemize{
\item \code{group}: grouping or conditions, e.g. "control" or "tumor",
required for most downstream analyses
\item \code{batch}: batch information, required for batch effect correction
}
}

\section{Aggregation}{
pGlyco3 performs quantification on the PSM level.
This level of information is too detailed for most downstream analyses.
This function aggregate PSMs into glycopeptides through summation.
For each glycopeptide (unique combination of "peptide", "peptide_site", "protein", "protein_site",
"gene", "glycan_composition", "glycan_structure"),
we sum up the quantifications of all PSMs that belong to this glycopeptide.
}

\seealso{
\code{\link[glyexp:experiment]{glyexp::experiment()}}, \code{\link[glyrepr:glycan_composition]{glyrepr::glycan_composition()}}
}
