% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/byonic-pglycoquant.R
\name{read_byonic_pglycoquant}
\alias{read_byonic_pglycoquant}
\title{Read Byonic-pGlycoQuant result}
\usage{
read_byonic_pglycoquant(
  fp,
  sample_info = NULL,
  quant_method = c("label-free", "TMT"),
  glycan_type = c("N", "O"),
  sample_name_converter = NULL,
  bitr_db = "org.Hs.eg.db"
)
}
\arguments{
\item{fp}{File path of the pGlycoQuant result file.}

\item{sample_info}{File path of the sample information file (csv),
or a sample information data.frame/tibble.}

\item{quant_method}{Quantification method. Either "label-free" or "TMT".}

\item{glycan_type}{Glycan type. Either "N" or "O". Default is "N".}

\item{sample_name_converter}{A function to convert sample names from file paths.
The function should take a character vector of old sample names
and return new sample names.
Note that sample names in \code{sample_info} should match the new names.
If NULL, original names are kept.}

\item{bitr_db}{name of the OrgDb package to use for UniProt to gene symbol conversion.
Default is "org.Hs.eg.db".}
}
\value{
An \code{\link[glyexp:experiment]{glyexp::experiment()}} object.
}
\description{
If you used Byonic for intact glycopeptide identification,
and used pGlycoQuant for quantification, this is the function for you.
It reads in a pGlycoQuant result file and returns a \code{\link[glyexp:experiment]{glyexp::experiment()}} object.
Currently only label-free quantification is supported.
}
\section{Which file to use?}{

You should use the "Quant.spectra.list" file in the pGlycoQuant result folder.
Files from Byonic result folder are not needed.
For instructions on how to use Byonic and pGlycoQuant, please refer to
the manual: \href{https://github.com/Power-Quant/pGlycoQuant/blob/main/Manual\%20for\%20pGlycoQuant_v202211.pdf}{pGlycoQuant}.
}

\section{Multisite glycopeptides}{

Currently, only single-site glycopeptides are supported.
Multisite glycopeptides will be removed.
}

\section{Output}{

This function returns a \code{\link[glyexp:experiment]{glyexp::experiment()}} object.

The following columns could be found in the variable information tibble:
\itemize{
\item \code{peptide}: character, peptide sequence
\item \code{glycan_composition}: \code{\link[glyrepr:glycan_composition]{glyrepr::glycan_composition()}}, glycan compositions.
\item \code{peptide_site}: integer, site of glycosylation on peptide
\item \code{protein}: character, protein accessions
\item \code{protein_site}: integer, site of glycosylation on protein
\item \code{gene}: character, gene symbols (if clusterProfiler package is available)
}
}

\section{Sample information}{
The sample information file should be a \code{csv} file with the first column
named \code{sample}, and the rest of the columns being sample information.
The \code{sample} column must match the \code{RawName} column in the pGlyco3 result file,
although the order can be different.

You can put any useful information in the sample information file.
Recommended columns are:
\itemize{
\item \code{group}: grouping or conditions, e.g. "control" or "tumor",
required for most downstream analyses
\item \code{batch}: batch information, required for batch effect correction
}
}

\seealso{
\code{\link[glyexp:experiment]{glyexp::experiment()}}, \code{\link[glyrepr:glycan_composition]{glyrepr::glycan_composition()}}
}
