% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glycan-finder.R
\name{read_glycan_finder}
\alias{read_glycan_finder}
\title{Read GlycanFinder result}
\usage{
read_glycan_finder(
  fp,
  sample_info = NULL,
  quant_method = "label-free",
  glycan_type = "N",
  sample_name_converter = NULL,
  parse_structure = TRUE
)
}
\arguments{
\item{fp}{File path of the GlycanFinder result CSV file.}

\item{sample_info}{File path of sample info or NULL for default.}

\item{quant_method}{Quantification method (currently only "label-free").}

\item{glycan_type}{Glycan type: "N", "O-GalNAc", "O-GlcNAc", etc.}

\item{sample_name_converter}{Function to convert sample names.}

\item{parse_structure}{Whether to parse structures.}
}
\value{
glyexp::experiment object
}
\description{
Peaks GlycanFinder is software for intact glycopeptide identification.
This function reads the result file and returns a \code{\link[glyexp:experiment]{glyexp::experiment()}} object.
}
