% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glyco-decipher.R
\name{read_glyco_decipher}
\alias{read_glyco_decipher}
\title{Read glyco-decipher output}
\usage{
read_glyco_decipher(
  fp,
  sample_info = NULL,
  quant_method = c("label-free", "TMT"),
  glycan_type = c("N", "O"),
  sample_name_converter = NULL,
  orgdb = "org.Hs.eg.db"
)
}
\arguments{
\item{fp}{File path of the pGlyco3 result file.}

\item{sample_info}{File path of the sample information file (csv),
or a sample information data.frame/tibble.}

\item{quant_method}{Quantification method. Either "label-free" or "TMT".}

\item{glycan_type}{Glycan type. Either "N" or "O". Default is "N".}

\item{sample_name_converter}{A function to convert sample names from file paths.
The function should take a character vector of old sample names
and return new sample names.
Note that sample names in \code{sample_info} should match the new names.
If NULL, original names are kept.}

\item{orgdb}{name of the OrgDb package to use for UniProt to gene symbol conversion.
Default is "org.Hs.eg.db".}
}
\value{
An \code{\link[glyexp:experiment]{glyexp::experiment()}} object.
}
\description{
Glyco-Decipher is a software for glycopeptide identification and quantification.
This function reads in the result file and returns a \code{\link[glyexp:experiment]{glyexp::experiment()}} object.
Currently only label-free quantification is supported.
}
\section{Which file to use?}{

You should use the "site.csv" file in the result folder.
This file contains "Site" and "Glycan" columns,
followed by quantification result for each sample.
}

\section{Protein inference and uncertain sites}{

Glyco-Decipher reports uncertain sites and proteins.
This function automatically performs protein inference using the
parsimony method to find the leader proteins.
This ensures each glycopeptide is uniquely mapped to a single protein, gene, and glycosite.
Besides, uncertain sites on proteins are assigned \code{NA}.
}

\section{Variable information}{


The following columns could be found in the variable information tibble:
\itemize{
\item \code{protein}: character, protein accession (after protein inference)
\item \code{protein_site}: integer, site of glycosylation on protein (after protein inference)
\item \code{gene}: character, gene name (symbol) (after protein inference)
\item \code{glycan_composition}: \code{\link[glyrepr:glycan_composition]{glyrepr::glycan_composition()}}, glycan compositions.
}
}

\section{Sample information}{
The sample information file should be a \code{csv} file with the first column
named \code{sample}, and the rest of the columns being sample information.
The \code{sample} column must match the \code{RawName} column in the pGlyco3 result file,
although the order can be different.

You can put any useful information in the sample information file.
Recommended columns are:
\itemize{
\item \code{group}: grouping or conditions, e.g. "control" or "tumor",
required for most downstream analyses
\item \code{batch}: batch information, required for batch effect correction
}
}

