% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glyhunter.R
\name{read_glyhunter}
\alias{read_glyhunter}
\title{Read GlyHunter result}
\usage{
read_glyhunter(
  fp,
  sample_info = NULL,
  glycan_type = "N",
  sample_name_converter = NULL
)
}
\arguments{
\item{fp}{File path of the GlyHunter result file.}

\item{sample_info}{File path of the sample information file (csv),
or a sample information data.frame/tibble.}

\item{glycan_type}{Glycan type. One of "N", "O-GalNAc", "O-GlcNAc", "O-Man", "O-Fuc", or "O-Glc".
Default is "N".}

\item{sample_name_converter}{A function to convert sample names from file paths.
The function should take a character vector of old sample names
and return new sample names.
Note that sample names in \code{sample_info} should match the new names.
If NULL, original names are kept.}
}
\value{
An \code{\link[glyexp:experiment]{glyexp::experiment()}} object.
}
\description{
This function reads in a \href{https://github.com/fubin1999/glyhunter}{GlyHunter}
result file and returns a \code{\link[glyexp:experiment]{glyexp::experiment()}} object.
}
\section{Which file to use?}{
Use the "summary_area.csv" file in the GlyHunter result folder.
No edit is necessary.
}

\section{Variable information}{
The following columns could be found in the variable information tibble:
\itemize{
\item \code{glycan_composition}: \code{\link[glyrepr:glycan_composition]{glyrepr::glycan_composition()}}, glycan compositions.
}
}

\section{Sample information}{
The sample information file should be a \code{csv} file with the first column
named \code{sample}, and the rest of the columns being sample information.
The \code{sample} column must match the \code{RawName} column in the pGlyco3 result file,
although the order can be different.

You can put any useful information in the sample information file.
Recommended columns are:
\itemize{
\item \code{group}: grouping or conditions, e.g. "control" or "tumor",
required for most downstream analyses
\item \code{batch}: batch information, required for batch effect correction
}
}

\seealso{
\code{\link[glyexp:experiment]{glyexp::experiment()}}, \code{\link[glyrepr:glycan_composition]{glyrepr::glycan_composition()}}
}
