% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pglyco3.R
\name{read_pglyco3}
\alias{read_pglyco3}
\title{Read pGlyco3 result}
\usage{
read_pglyco3(
  fp,
  sample_info = NULL,
  quant_method = c("label-free", "TMT"),
  glycan_type = c("N", "O"),
  sample_name_converter = NULL,
  parse_structure = FALSE
)
}
\arguments{
\item{fp}{File path of the pGlyco3 result file.}

\item{sample_info}{File path of the sample information file (csv),
or a sample information data.frame/tibble.}

\item{quant_method}{Quantification method. Either "label-free" or "TMT".}

\item{glycan_type}{Glycan type. Either "N" or "O". Default is "N".}

\item{sample_name_converter}{A function to convert sample names from file paths.
The function should take a character vector of old sample names
and return new sample names.
Note that sample names in \code{sample_info} should match the new names.
If NULL, original names are kept.}

\item{parse_structure}{Logical. Whether to parse glycan structures.
If \code{TRUE}, glycan structures are parsed and included in the
\code{var_info} as \code{glycan_structure} column. If \code{FALSE} (default), structure parsing
is skipped and structure-related columns are removed.}
}
\value{
An \code{\link[glyexp:experiment]{glyexp::experiment()}} object.
}
\description{
pGlyco3 is a software for intact glycopeptide identification and quantification.
This function reads in the result file and returns a \code{\link[glyexp:experiment]{glyexp::experiment()}} object.
Currently only label-free quantification is supported.
}
\section{Which file to use?}{
You should use the result file from pGlyco3 that contains quantification information.
The file should have columns including \code{RawName}, \code{MonoArea}, \code{Peptide}, \code{Proteins},
\code{Genes}, \code{GlycanComposition}, \code{PlausibleStruct}, \code{GlySite}, and \code{ProSites}.
}

\section{Sample information}{
The sample information file should be a \code{csv} file with the first column
named \code{sample}, and the rest of the columns being sample information.
The \code{sample} column must match the \code{RawName} column in the pGlyco3 result file,
although the order can be different.

You can put any useful information in the sample information file.
Recommended columns are:
\itemize{
\item \code{group}: grouping or conditions, e.g. "control" or "tumor",
required for most downstream analyses
\item \code{batch}: batch information, required for batch effect correction
}
}

\section{Protein inference}{
pGlyco3 reports protein groups.
That is, shared glycopeptides are reported as a group of proteins separated by ";".
This function automatically performs protein inference using the
parsimony method to find the leader proteins.
This ensures each glycopeptide is uniquely mapped to a single protein, gene, and glycosite.
}

\section{Aggregation}{
pGlyco3 performs quantification on the PSM level.
This level of information is too detailed for most downstream analyses.
This function aggregate PSMs into glycopeptides through summation.
For each glycopeptide (unique combination of "peptide", "peptide_site", "protein", "protein_site",
"gene", "glycan_composition", "glycan_structure"),
we sum up the quantifications of all PSMs that belong to this glycopeptide.
}

\section{Variable information}{
The following columns could be found in the variable information tibble:
\itemize{
\item \code{peptide}: character, peptide sequence
\item \code{peptide_site}: integer, site of glycosylation on peptide
\item \code{protein}: character, protein accession (after protein inference)
\item \code{protein_site}: integer, site of glycosylation on protein (after protein inference)
\item \code{gene}: character, gene name (symbol) (after protein inference)
\item \code{glycan_composition}: \code{\link[glyrepr:glycan_composition]{glyrepr::glycan_composition()}}, glycan compositions.
\item \code{glycan_structure}: \code{\link[glyrepr:glycan_structure]{glyrepr::glycan_structure()}}, glycan structures (if \code{parse_structure = TRUE}).
}
}

\section{Glycan structures}{
pGlyco3 reports a "plausible structure" for each glycan.
You can set \code{parse_structure = TRUE} to parse these structures into a "glycan_structure"
column as a \code{\link[glyrepr:glycan_structure]{glyrepr::glycan_structure()}} vector.
However, please take caution with these structures,
because pGlyco3 does not have strict quality control on glycan structure annotations.
}

\seealso{
\code{\link[glyexp:experiment]{glyexp::experiment()}}, \code{\link[glyrepr:glycan_composition]{glyrepr::glycan_composition()}},
\code{\link[glyrepr:glycan_structure]{glyrepr::glycan_structure()}}
}
