% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glyread-package.R
\docType{package}
\name{glyread-package}
\alias{glyread}
\alias{glyread-package}
\title{glyread: Read and process glycomics and glycoproteomics data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a unified interface for reading and processing quantification results from various glycomics and glycoproteomics software tools, including Byonic (ByoLogic, pGlycoQuant), StrucGP, pGlyco3 (pGlycoQuant), Glyco-Decipher, GlyHunter, and MSFragger. The package standardizes raw data from different sources, performs peptide-spectrum match (PSM) aggregation, parses glycan compositions and structures, and converts the data into a standardized experiment object from the 'glyexp' package for downstream analysis within the glycoverse ecosystem.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://glycoverse.github.io/glyread/}
}

}
\author{
\strong{Maintainer}: Bin Fu \email{23110220018@m.fudan.edu.cn} (\href{https://orcid.org/0000-0001-8567-2997}{ORCID})

}
\keyword{internal}
