% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msfragger.R
\name{read_msfragger}
\alias{read_msfragger}
\title{Read MSFragger-Glyco result}
\usage{
read_msfragger(
  dp,
  sample_info = NULL,
  quant_method = "label-free",
  glycan_type = "N",
  sample_name_converter = NULL
)
}
\arguments{
\item{dp}{The directory path of the MSFragger-Glyco result folder.}

\item{sample_info}{File path of the sample information file (csv),
or a sample information data.frame/tibble.}

\item{quant_method}{Quantification method. Either "label-free" or "TMT".}

\item{glycan_type}{Glycan type. Either "N" or "O". Default is "N".}

\item{sample_name_converter}{A function to convert sample names from file paths.
The function should take a character vector of old sample names
and return new sample names.
Note that sample names in \code{sample_info} should match the new names.
If NULL, original names are kept.}
}
\value{
An \code{\link[glyexp:experiment]{glyexp::experiment()}} object.
}
\description{
MSFragger-Glyco is a software for glycopeptide identification and quantification.
This function reads in the result file and returns a \code{\link[glyexp:experiment]{glyexp::experiment()}} object.
}
\details{
This function uses the "psm.tsv" file in each sample folder.
Sample names are extracted from the file paths.
They are the parent of each "psm.tsv" file.
For example, "msfragger_result/H1/psm.tsv" will be named "H1".
}
\section{Variable information}{

The following columns could be found in the variable information tibble:
\itemize{
\item \code{peptide}: character, peptide sequence
\item \code{peptide_site}: integer, site of glycosylation on peptide
\item \code{protein}: character, protein accession
\item \code{protein_site}: integer, site of glycosylation on protein
\item \code{gene}: character, gene name (symbol)
\item \code{glycan_composition}: \code{\link[glyrepr:glycan_composition]{glyrepr::glycan_composition()}}, glycan compositions.
}
}

\seealso{
\code{\link[glyexp:experiment]{glyexp::experiment()}}, \code{\link[glyrepr:glycan_composition]{glyrepr::glycan_composition()}}
}
