% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pglyco3.R
\name{read_pglyco3}
\alias{read_pglyco3}
\title{Read pGlyco3 Result}
\usage{
read_pglyco3(
  fp,
  sample_info_fp = NULL,
  name = NULL,
  quantify_on = c("mono", "sum"),
  differ_a_g = FALSE,
  parse_structure = TRUE,
  describe_glycans = TRUE
)
}
\arguments{
\item{fp}{File path of the pGlyco3 result file.}

\item{sample_info_fp}{File path of the sample information file.}

\item{name}{Name of the experiment. If not provided, a default name with
current time will be used.}

\item{quantify_on}{Quantify on "mono" or "sum". If "mono", the "MonoArea"
column will be used for quantification. If "sum", the "IsotopeArea" column
will be used for quantification. Default is "mono".}

\item{differ_a_g}{Whether to differentiate "A" (Neu5Ac) and "G" (Neu5Gc).
If \code{FALSE}, "A" and "G" will be replaced by "S" in compositions and structures.
Default is \code{FALSE}.}

\item{parse_structure}{Whether to parse glycan structures. Default is \code{TRUE}.}

\item{describe_glycans}{Whether to describe glycan properties. Default is \code{TRUE}.
If \code{parse_structure} is \code{FALSE}, this argument will be forced to \code{FALSE}.}
}
\value{
An \code{\link[glyexp:experiment]{glyexp::experiment()}} object.
}
\description{
This function reads pGlyco3 quantification result,
and returns an \code{\link[glyexp:experiment]{glyexp::experiment()}} object.
}
\section{Input}{
You should use the \code{txt} file in pGlyco3 result folder with "Quant" in the
file name. It is something like \code{pGlycoDB-GP-FDR-Pro-Quant.txt}.

The sample information file should be a \code{csv} file with the first column
named \code{sample}, and the rest of the columns are sample information.
The \code{sample} column must match the \code{RawName} column in the pGlyco3 result file,
although the order can be different.
You can put any useful information in the sample information file.
Recommended columns are:
\itemize{
\item \code{group}: grouping or conditions, e.g. "control" or "tumor",
required for most downstream analyses
\item \code{batch}: batch information, required for batch effect correction
\item \code{bio_replicate}:  Indicates the identity of biologically distinct samples.
While the \code{sample} column is typically used as a unique identifier for each run,
\code{bio_replicate} specifies whether the samples originate from the same biological source.
In common experimental designs,
multiple technical replicates are often derived from the same biological sample,
making the bio_replicate column essential for distinguishing
biological replicates from technical replicates.
}
}

\section{Output}{
The result is an \code{\link[glyexp:experiment]{glyexp::experiment()}} object.
If \code{parse_structure} is \code{TRUE} (by default),
a list of parsed \code{glycan_graph} objects will be stored in the experiment,
with the glycan structure strings as names.
It can be accessed by \code{exp$glycan_graphs} or \code{exp[["glycan_graphs"]]}.

The following columns are in the variable information tibble:
\itemize{
\item \code{charge}: integer, charge state
\item \code{peptide}: character, peptide sequence
\item \code{modifications}: character, modifications other than glycosylation,
separated by semicolon, e.g. \verb{5,Carbamidomethyl[C];10,Carbamidomethyl[C]}
\item \code{glycan_composition}: character, glycan composition, e.g. "H(5)N(4)A(1)F(1)"
\item \code{glycan_structure}: character, pGlyco-style structure strings, renamed from
\code{PlausibleStruct} in the original file
\item \code{n_hex}: integer, number of Hex
\item \code{n_hexnac}: integer, number of HexNAc
\item \code{n_neuac}: integer, number of NeuAc
\item \code{n_fuc}: integer, number of Fuc (dHex)
\item \code{peptide_site}: integer, site of glycosylation on peptide
\item \code{proteins}: character, protein names, separated by semicolon
\item \code{genes}: character, gene names, separated by semicolon
\item \code{protein_sites}: character, site of glycosylation on protein,
separated by semicolon
}

If \code{describe_glycans} is \code{TRUE} (by default),
additional columns will be added to the variable information tibble:
\itemize{
\item \code{glycan_type}: character, N-glycan type, either "complex", "highmannose",
"hybrid", or "paucimannose"
\item \code{bisecting}: logical, whether the glycan has a bisecting GlcNAc
\item \code{n_antennae}: integer, number of antennae
\item \code{n_core_fuc}: integer, number of core fucose
\item \code{n_arm_fuc}: integer, number of arm fucose
\item \code{n_gal}: integer, number of Gal
\item \code{n_terminal_gal}: integer, number of terminal Gal
}

Besides, glycan compositions are reformatted into condensed format,
e.g. "H5N4F1S1" for "H(5)N(4)A(1)F(1)".
Note that "A" and "G" in glycan structures are replaced by "S"
if \code{differ_a_g} is \code{FALSE}.
}

