% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strucgp.R
\name{read_strucgp}
\alias{read_strucgp}
\title{Read the result from StrucGP}
\usage{
read_strucgp(fp)
}
\arguments{
\item{fp}{File path of the StrucGP result file.}
}
\value{
A tibble with the processed variable information.
}
\description{
StrucGP is a software for intact glycopeptide identification.
As StrucGP doesn't support quantification,
this function only returns the processed variable information tibble,
instead of an \code{\link[glyexp:experiment]{glyexp::experiment()}} object like other \code{glyread} functions.
}
